refactor: removed database implementation from the app
